songlist container refactor
